remove div and recoil state, add types
